adding in 3d layer support
I've started with the 3d layer support. I think the best way is to define a new base layer, and parse the network based on the dimensionality of the network, which we can have as a command line flag. Let's have a initial goal for just parsing a sequential snippet of the x3d_m model (no optimisations or anything). Then we can look at better coverage.